Modification des types dans LogModel et StatusJobModel.

+++ EasySaveLib/Models/LogModel.cs
@@ -14,10 +14,10 @@ namespace EasySaveLib.Models
         public ulong Timestamp { get; set; }
         public string Source { get; set; }
         public string Destination { get; set; }
-        public int Size { get; set; }
+        public ulong Size { get; set; }
         public float TransfertTime { get; set; }
         
-        public LogModel(string name, ulong timestamp, string source, string destination, int size, float transfertTime)
+        public LogModel(string name, ulong timestamp, string source, string destination, ulong size, float transfertTime)
         {
             Name = name;
             Timestamp = timestamp;
